+++ KeeAnywhere/PluginResources.Designer.cs
@@ -63,6 +63,16 @@ namespace KeeAnywhere {
         /// <summary>
         ///   Looks up a localized resource of type System.Drawing.Bitmap.
         /// </summary>
+        internal static System.Drawing.Bitmap AmazonDrive_16x16 {
+            get {
+                object obj = ResourceManager.GetObject("AmazonDrive_16x16", resourceCulture);
+                return ((System.Drawing.Bitmap)(obj));
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized resource of type System.Drawing.Bitmap.
+        /// </summary>
         internal static System.Drawing.Bitmap Dropbox_16x16 {
             get {
                 object obj = ResourceManager.GetObject("Dropbox_16x16", resourceCulture);