+++ KeeAnywhere/StorageProviders/StorageRegistry.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using KeeAnywhere.StorageProviders.AmazonDrive;
 using KeeAnywhere.StorageProviders.Dropbox;
 using KeeAnywhere.StorageProviders.GoogleDrive;
 using KeeAnywhere.StorageProviders.HubiC;
@@ -11,6 +12,7 @@ namespace KeeAnywhere.StorageProviders
         public static IEnumerable<StorageDescriptor> Descriptors = new[]
         {
             // Preserve sort order: ascending!
+            new StorageDescriptor(StorageType.AmazonDrive, "Amazon Drive", "adrive", account => new AmazonDriveStorageProvider(account), () => new AmazonDriveStorageConfigurator(), PluginResources.AmazonDrive_16x16),
             new StorageDescriptor(StorageType.Dropbox, "Dropbox", "dropbox", account => new DropboxStorageProvider(account), () => new DropboxStorageConfigurator(false), PluginResources.Dropbox_16x16),
             new StorageDescriptor(StorageType.DropboxRestricted, "Dropbox-Restricted", "dropbox-r", account => new DropboxStorageProvider(account), () => new DropboxStorageConfigurator(true), PluginResources.Dropbox_16x16),
             new StorageDescriptor(StorageType.GoogleDrive, "Google Drive", "gdrive", account => new GoogleDriveStorageProvider(account), () => new GoogleDriveStorageConfigurator(), PluginResources.GoogleDrive_16x16),