+++ KeeAnywhere/StorageProviders/StorageType.cs
@@ -3,6 +3,7 @@ namespace KeeAnywhere.StorageProviders
     public enum StorageType
     {
         // Preserve sort order: ascending!
+        AmazonDrive,
         Dropbox,
         DropboxRestricted,
         GoogleDrive,